Renamed exceptions to errors
