+++ packages/common/src/errors.js
