+++ packages/gatekeeper/src/db-redis.js
@@ -1,5 +1,5 @@
 import Redis from 'ioredis';
-import { InvalidDIDError } from '@mdip/common/exceptions';
+import { InvalidDIDError } from '@mdip/common/errors';
 
 let redis;
 let DB;
